third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-transform-function-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <transform-function> with additivity
  assert_equals: expected "translateX(350px)" but got "translate(350px)"
[FAIL] Animating a custom property of type <transform-function> with a single keyframe and additivity
  assert_equals: expected "translateX(250px)" but got "translate(250px)"
[FAIL] Animating a custom property of type <transform-function> with iterationComposite
  assert_equals: expected "translateX(250px)" but got "translateX(50px)"
Harness: the test ran to completion.
